requirements.txt
attrs==25.3.0
cattrs==25.1.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
fastf1==3.6.0
fonttools==4.59.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
RapidFuzz==3.13.0
requests==2.32.4
requests-cache==1.2.1
scipy==1.16.0
six==1.17.0
SQLAlchemy==2.0.41
timple==0.1.8
typing_extensions==4.14.1
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
websockets==13.1
